Embed Preview: do not use SandBox
SandBox is meant to tightly scope embed side effects so that they can't
break the rest of the editor, but since we're already just sending an
iframe tag, we know it will be contained.

Additionally, the SandBox interferes with ActBlue's ability to discern
the hosting domain, so it will block embeds loaded from SandBoxes.